LayoutTests/fast/dom/CSSStyleDeclaration/css-properties-case-sensitive-expected.txt

This test checks that access to CSS properties via JavaScript properties on DOM elements is case sensitive.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


normal cases

PASS element.style.zIndex is '1'
PASS element.style.ZIndex is undefined.

"css" prefix

PASS element.style.cssZIndex is '1'
PASS element.style.CssZIndex is '1'
PASS element.style.CsszIndex is undefined.
PASS element.style.csszIndex is undefined.
PASS successfullyParsed is true

TEST COMPLETE
